requirements.txt
Django>=4.2.7
django-filter>=23.3
djangorestframework>=3.14.0
whitenoise>=6.6.0